Refactor system to do linting.
